update ci.yml and release.yml to work together
This is a bit unfortunate as we push up a zip file with the "zip we want" and a "sha1" in it. This is not usable by grafana, but our release process now unpacks this zip file and puts them on the release, so that the necessary files are present there.